incubators/backend/requirements.txt
fastapi[all]
uvicorn[standard]
asyncpg
alembic
sqlmodel
aiomqtt
redis
aiofiles
python-multipart
python-jose[cryptography]
passlib[bcrypt]
email-validator
pydantic-settings
minio
httpx
pytest
pytest-asyncio
google-genai
